chore: update vue-mgl

+++ vue-mgl-tools/update-changelog.sh
@@ -3,8 +3,12 @@
 set -x
 set -e
 
-rm ./CHANGELOG.en-US.md
-rm ./CHANGELOG.zh-CN.md
+if [ -f ./CHANGELOG.en-US.md ]; then
+  rm ./CHANGELOG.en-US.md
+fi
+if [ -f ./CHANGELOG.zh-CN.md ]; then
+  rm ./CHANGELOG.zh-CN.md
+fi
 
 wget https://cdn.jsdelivr.net/npm/vue-mgl@latest/CHANGELOG.md -O CHANGELOG.en-US.md
 wget https://cdn.jsdelivr.net/npm/vue-mgl@latest/CHANGELOG.md -O CHANGELOG.zh-CN.md
